client/requirements.txt
psutil==5.8.0
Flask==1.0.2
colorclass==2.2.0
requests==2.25.1
terminaltables==3.1.0
gevent_socketio==0.3.6
